requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic
pydantic-settings

# Notion API
notion-client==2.2.1

# OpenAI
openai==1.6.1
tiktoken

# Vector Database - Using FAISS for Python 3.14 compatibility
faiss-cpu
# chromadb  # Temporarily disabled for Python 3.14
# qdrant-client==1.7.0  # Optional for production

# Text Processing
# langchain==0.1.0
# langchain-community==0.2.0
# langchain-openai==0.0.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database (for metadata)
sqlalchemy==2.0.23
alembic==1.13.0
# asyncpg==0.29.0  # For PostgreSQL (optional)

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Utils
python-dateutil==2.8.2
pyyaml==6.0.1